RF: Use the same StdOut panel for Runner and Coder so they get the same buttons and styling
