+++ psychopy/app/stdout/__init__.py
@@ -1,0 +1,1 @@
+from .stdOutRich import StdOutRich, ScriptOutputPanel, ScriptOutputCtrl
